+++ libfaac/frame.c
@@ -51,8 +51,8 @@ static SR_INFO srInfo[12+1];
 
 // default bandwidth/samplerate ratio
 static const struct {
-    double fac;
-    double freq;
+    float fac;
+    float freq;
 } g_bw = {0.42, 18000};
 
 int FAACAPI faacEncGetVersion( char **faac_id_string,
@@ -157,13 +157,13 @@ int FAACAPI faacEncSetConfiguration(faacEncHandle hpEncoder,
 
     if (config->bitRate && !config->bandWidth)
     {
-        config->bandWidth = (double)config->bitRate * hEncoder->sampleRate * g_bw.fac / 50000.0;
+        config->bandWidth = (float)config->bitRate * hEncoder->sampleRate * g_bw.fac / 50000.0;
         if (config->bandWidth > g_bw.freq)
             config->bandWidth = g_bw.freq;
 
         if (!config->quantqual)
         {
-            config->quantqual = (double)config->bitRate * hEncoder->numChannels / 1280;
+            config->quantqual = (float)config->bitRate * hEncoder->numChannels / 1280;
             if (config->quantqual > 100)
                 config->quantqual = (config->quantqual - 100) * 3.0 + 100;
         }
@@ -358,7 +358,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
     BitStream *bitStream; /* bitstream used for writing the frame to */
 #ifdef DRM
     int desbits, diff;
-    double fix;
+    float fix;
 #endif
 
     /* local copy's of parameters */
@@ -388,11 +388,11 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
     /* Update current sample buffers */
     for (channel = 0; channel < numChannels; channel++)
 	{
-		double *tmp;
+		float *tmp;
 
 
 		if (!hEncoder->sampleBuff[channel])
-			hEncoder->sampleBuff[channel] = (double*)AllocMemory(FRAME_LEN*sizeof(double));
+			hEncoder->sampleBuff[channel] = (float*)AllocMemory(FRAME_LEN*sizeof(float));
 
 		tmp = hEncoder->sampleBuff[channel];
 
@@ -418,7 +418,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
 
 						for (i = 0; i < samples_per_channel; i++)
 						{
-							hEncoder->next3SampleBuff[channel][i] = (double)*input_channel;
+							hEncoder->next3SampleBuff[channel][i] = (float)*input_channel;
 							input_channel += numChannels;
 						}
 					}
@@ -430,7 +430,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
 
 						for (i = 0; i < samples_per_channel; i++)
 						{
-							hEncoder->next3SampleBuff[channel][i] = (1.0/256) * (double)*input_channel;
+							hEncoder->next3SampleBuff[channel][i] = (1.0/256) * (float)*input_channel;
 							input_channel += numChannels;
 						}
 					}
@@ -442,7 +442,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
 
 						for (i = 0; i < samples_per_channel; i++)
 						{
-							hEncoder->next3SampleBuff[channel][i] = (double)*input_channel;
+							hEncoder->next3SampleBuff[channel][i] = (float)*input_channel;
 							input_channel += numChannels;
 						}
 					}
@@ -480,7 +480,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
     hEncoder->psymodel->PsyCalculate(channelInfo, &hEncoder->gpsyInfo, hEncoder->psyInfo,
         hEncoder->srInfo->cb_width_long, hEncoder->srInfo->num_cb_long,
         hEncoder->srInfo->cb_width_short,
-        hEncoder->srInfo->num_cb_short, numChannels, (double)hEncoder->aacquantCfg.quality / DEFQUAL);
+        hEncoder->srInfo->num_cb_short, numChannels, (float)hEncoder->aacquantCfg.quality / DEFQUAL);
 
     hEncoder->psymodel->BlockSwitch(coderInfo, hEncoder->psyInfo, numChannels);
 
@@ -546,7 +546,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
 	}
 
     AACstereo(coderInfo, channelInfo, hEncoder->freqBuff, numChannels,
-              (double)hEncoder->aacquantCfg.quality/DEFQUAL, jointmode);
+              (float)hEncoder->aacquantCfg.quality/DEFQUAL, jointmode);
 
 #ifdef DRM
     /* loop the quantization until the desired bit-rate is reached */
@@ -568,13 +568,13 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
     frameBytes = CloseBitStream(bitStream);
 
     /* now calculate desired bits and compare with actual encoded bits */
-    desbits = (int) ((double) numChannels * (hEncoder->config.bitRate * FRAME_LEN)
+    desbits = (int) ((float) numChannels * (hEncoder->config.bitRate * FRAME_LEN)
             / hEncoder->sampleRate);
 
     diff = ((frameBytes - 1 /* CRC */) * 8) - desbits;
 
     /* do linear correction according to relative difference */
-    fix = (double) desbits / ((frameBytes - 1 /* CRC */) * 8);
+    fix = (float) desbits / ((frameBytes - 1 /* CRC */) * 8);
 
     /* speed up convergence. A value of 0.92 gives approx up to 10 iterations */
     if (fix > 0.92)
@@ -618,7 +618,7 @@ int FAACAPI faacEncEncode(faacEncHandle hpEncoder,
     {
         int desbits = numChannels * (hEncoder->config.bitRate * FRAME_LEN)
             / hEncoder->sampleRate;
-        double fix = (double)desbits / (double)(frameBytes * 8);
+        float fix = (float)desbits / (float)(frameBytes * 8);
 
         if (fix < 0.9)
             fix += 0.1;
